added some common vector operations

+++ include/core/src/Game.cpp
@@ -53,8 +53,8 @@ void Game::render() {
 
 // locks framerate to a fixed value defined by 'FPS'
 void Game::delay(std::chrono::nanoseconds frameTimeElapsed) {
-  int frameTicks = frameTimeElapsed.count();
-  const int SCREEN_TICKS_PER_FRAME = 1000 / static_cast<float>(FPS);
+  int frameTicks = static_cast<int>(frameTimeElapsed.count());
+  const int SCREEN_TICKS_PER_FRAME = static_cast<int>(1000 / static_cast<float>(FPS));
 
   if (frameTicks < SCREEN_TICKS_PER_FRAME) {
     SDL_Delay(SCREEN_TICKS_PER_FRAME - frameTicks); // wait remaining time
